+++ Model/Attribute/InputType/Text.php
@@ -13,12 +13,4 @@ namespace Alekseon\AlekseonEav\Model\Attribute\InputType;
 class Text extends AbstractInputType
 {
     protected $defaultBackendType = 'varchar';
-
-    /**
-     * @return bool
-     */
-    public function canUseInputValidator()
-    {
-        return true;
-    }
 }